Make tests work without `mypy` installed
Move the `mypy.api` import into the test that is going to use it,
rather than doing it in the global scope.  This makes it possible to run
the remainder of the test suite without `mypy` installed.  This is
important for users who wish to run unit tests to verify that
the package works correctly in production setup, when `mypy`
is not a necessary or useful dependency for production systems.